# docs/swagger.yaml
basePath: /gateway/v1
definitions:
...
  dto.ChangeGroupMemberRequest:
    properties:
      gid:
        example: av8FMdRdcb
        type: string
      uids:
        example:
        - av8FMdRdcb
        - av8FMdRdcc
        items:
          type: string
        maxItems: 20
        minItems: 1
        type: array
    required:
    - gid
    - uids
    type: object
  dto.ChangeGroupMemberResponse:
    properties:
      count:
        example: 1
        type: integer
    type: object
...
  dto.CreateGroupRequest:
    properties:
      avatar:
        example: https://example.com/avatar.png
        type: string
      desc:
        example: test
        maxLength: 128
        type: string
      members:
        example:
        - av8FMdRdcb
        - av8FMdRdcc
        items:
          type: string
        maxItems: 20
        minItems: 2
        type: array
      name:
        example: test
        maxLength: 32
        type: string
    required:
    - members
    - name
    type: object
...
  dto.DeleteGroupRequest:
    properties:
      gid:
        example: av8FMdRdcb
        type: string
    required:
    - gid
    type: object
...
  dto.Group:
    properties:
      avatar:
        example: https://example.com/avatar.png
        type: string
      desc:
        example: test
        type: string
      gid:
        example: av8FMdRdcb
        type: string
      max_member:
        example: 20
        type: integer
      member_count:
        example: 2
        type: integer
      members:
        items:
          $ref: '#/definitions/dto.GroupMember'
        type: array
      name:
        example: test
        type: string
      owner:
        $ref: '#/definitions/dto.GroupMember'
      owner_uid:
        example: av8FMdRdcb
        type: string
      status:
        example: 1
        type: integer
    type: object
  dto.GroupMember:
    properties:
      gid:
        example: av8FMdRdcb
        type: string
      status:
        description: '0: normal, 1: silent'
        example: 1
        type: integer
      type:
        description: '0: owner, 1: member'
        example: 1
        type: integer
      uid:
        example: av8FMdRdcb
        type: string
      user:
        $ref: '#/definitions/dto.User'
        description: only when withMembers is true and withInfo is true
    type: object
  dto.Message:
    properties:
      content:
        example: hello
        type: string
      contentType:
        example: 1
        type: integer
      createTime:
        example: 1579098983
        type: integer
      from:
        example: av8FMdRdcb
        type: string
      messageId:
        example: 1
        type: integer
      sessionId:
        example: "1"
        type: string
      sessionType:
        example: 1
        type: integer
      to:
        example: av8FMdRdcb
        type: string
    type: object
  dto.SendMessageReq:
    properties:
      content:
        example: hello
        type: string
      contentType:
        example: 1
        type: integer
      from:
        example: av8FMdRdcb
        type: string
      sessionId:
        example: "1"
        type: string
      sessionType:
        example: 1
        type: integer
      to:
        example: av8FMdRdcb
        type: string
    required:
    - content
    - contentType
    - from
    - sessionId
    - sessionType
    - to
    type: object
  dto.SendMessageResp:
    properties:
      messageId:
        example: 1
        type: integer
      sessionId:
        example: abc
        type: string
    type: object
  dto.UpdateGroupRequest:
    properties:
      avatar:
        example: https://www.example.com/avatar.png
        type: string
      desc:
        example: test
        maxLength: 128
        type: string
      gid:
        example: av8FMdRdcb
        type: string
      name:
        example: test
        maxLength: 32
        type: string
    required:
    - gid
    type: object
...
  dto.User:
    properties:
      avatar:
        example: https://www.example.com/avatar.png
        type: string
      connectUrl:
        example: ws://10.0.0.1:8080/ws
        type: string
      email:
        example: abc@example.com
        type: string
      loginStatus:
        example: 0
        type: integer
      name:
        example: user1
        type: string
      phone:
        example: "13800138000"
        type: string
      uid:
        example: av8FMdRdcb
        type: string
    type: object
...
  response.Meta:
    properties:
      extra:
        additionalProperties:
          type: string
        type: object
      page:
        example: 1
        minimum: 1
        type: integer
      pageSize:
        example: 10
        minimum: 1
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      reason:
        type: string
    type: object
info:
  contact:
    email: yusankurban@gmail.com
    name: Yusank
    url: https://yusank.space
  description: GoIM.Gateway 服务器接口文档
  license:
    name: MIT
    url: https://github.com/go-goim/core/blob/main/LICENSE
  termsOfService: http://go-goim.github.io/
  title: GoIM.Gateway Swagger
  version: "1.0"
paths:
...
  /group/create:
    post:
      consumes:
      - application/json
      description: 创建群组
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建群组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Group'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建群组
      tags:
      - 群组
  /group/delete:
    post:
      consumes:
      - application/json
      description: 删除群组
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除群组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除群组
      tags:
      - 群组
  /group/get:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取群组信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 群组ID
        in: query
        name: gid
        required: true
        type: string
      - description: 是否获取群组成员
        in: query
        name: with_members
        type: boolean
      - description: 是否获取群组信息
        in: query
        name: with_info
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Group'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取群组信息
      tags:
      - 群组
  /group/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取群组列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Group'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取群组列表
      tags:
      - 群组
  /group/member/invite:
    post:
      consumes:
      - application/json
      description: 任何群成员都可以添加群组成员
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 加入群组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeGroupMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ChangeGroupMemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 添加群组成员
      tags:
      - 群组
  /group/member/leave:
    post:
      consumes:
      - application/json
      description: 群组成员可以退出群组
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 退出群组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeGroupMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 退出群组
      tags:
      - 群组
  /group/member/remove:
    post:
      consumes:
      - application/json
      description: 群管理员可以删除群组成员
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除群组成员请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeGroupMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ChangeGroupMemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除群组成员
      tags:
      - 群组
  /group/update:
    post:
      consumes:
      - application/json
      description: 更新群组
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新群组请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Group'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新群组
      tags:
      - 群组
  /message/broadcast:
    post:
      consumes:
      - application/json
      description: 发送广播消息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SendMessageResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送广播消息
      tags:
      - message
  /message/send_msg:
    post:
      consumes:
      - application/json
      description: 发送单聊消息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SendMessageResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送单聊消息
      tags:
      - message
  /offline_message/query:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 查询离线消息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: lastMessageId
        in: query
        name: lastMessageId
        required: true
        type: integer
      - description: onlyCount
        in: query
        name: onlyCount
        type: boolean
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Message'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询离线消息
      tags:
      - offline_message
...
swagger: "2.0"